Pause menu fixed and level 1 cleaned up
The pause menu no longer activates during the countdown or the death screen. which fixes a lot of things. The pause button is also fixed so you don't have to press it twice to pause. level 1 camera is zoomed out more and the platform bars are lowered. Level 2 is starting.